requirements.txt
numpy>=1.26.0
pandas>=2.1.0
flask>=2.3.0
werkzeug>=3.0.0
google-auth>=2.3.0
google-auth-oauthlib>=0.4.6
google-api-python-client>=2.31.0
google-analytics-data>=0.17.1
python-dotenv>=0.19.1
